requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.22.99
numpy
Pillow
scipy>=0.17.0
